refactor(tui): share tree tool summaries
